Rename error generators
