+++ codegen-client/src/main/kotlin/software/amazon/smithy/rust/codegen/client/smithy/generators/ServiceGenerator.kt
@@ -14,7 +14,7 @@ import software.amazon.smithy.rust.codegen.client.smithy.generators.protocol.Pro
 import software.amazon.smithy.rust.codegen.core.rustlang.Attribute
 import software.amazon.smithy.rust.codegen.core.rustlang.RustModule
 import software.amazon.smithy.rust.codegen.core.smithy.RustCrate
-import software.amazon.smithy.rust.codegen.core.smithy.generators.error.TopLevelErrorGenerator
+import software.amazon.smithy.rust.codegen.core.smithy.generators.error.ServiceErrorGenerator
 import software.amazon.smithy.rust.codegen.core.smithy.generators.protocol.ProtocolSupport
 import software.amazon.smithy.rust.codegen.core.util.inputShape
 
@@ -56,7 +56,7 @@ class ServiceGenerator(
             }
         }
 
-        TopLevelErrorGenerator(clientCodegenContext, operations).render(rustCrate)
+        ServiceErrorGenerator(clientCodegenContext, operations).render(rustCrate)
 
         rustCrate.withModule(RustModule.Config) {
             ServiceConfigGenerator.withBaseBehavior(
