+++ codegen-core/src/main/kotlin/software/amazon/smithy/rust/codegen/core/smithy/generators/error/OperationErrorGenerator.kt
@@ -55,7 +55,7 @@ fun UnionShape.eventStreamErrorSymbol(symbolProvider: RustSymbolProvider): Runti
  * Generates a unified error enum for [operation]. [ErrorGenerator] handles generating the individual variants,
  * but we must still combine those variants into an enum covering all possible errors for a given operation.
  */
-class CombinedErrorGenerator(
+class OperationErrorGenerator(
     private val model: Model,
     private val symbolProvider: RustSymbolProvider,
     private val operationSymbol: Symbol,
