+++ codegen-core/src/test/kotlin/software/amazon/smithy/rust/codegen/core/smithy/generators/error/OperationErrorGeneratorTest.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.rust.codegen.server.smithy.generators
+package software.amazon.smithy.rust.codegen.core.smithy.generators.error
 
 import org.junit.jupiter.api.Test
 import software.amazon.smithy.model.shapes.StructureShape
@@ -12,12 +12,12 @@ import software.amazon.smithy.rust.codegen.core.smithy.transformers.OperationNor
 import software.amazon.smithy.rust.codegen.core.testutil.TestWorkspace
 import software.amazon.smithy.rust.codegen.core.testutil.asSmithyModel
 import software.amazon.smithy.rust.codegen.core.testutil.compileAndTest
+import software.amazon.smithy.rust.codegen.core.testutil.renderWithModelBuilder
+import software.amazon.smithy.rust.codegen.core.testutil.testSymbolProvider
 import software.amazon.smithy.rust.codegen.core.testutil.unitTest
 import software.amazon.smithy.rust.codegen.core.util.lookup
-import software.amazon.smithy.rust.codegen.server.smithy.testutil.serverRenderWithModelBuilder
-import software.amazon.smithy.rust.codegen.server.smithy.testutil.serverTestSymbolProvider
 
-class ServerCombinedErrorGeneratorTest {
+class OperationErrorGeneratorTest {
     private val baseModel = """
         namespace error
 
@@ -28,7 +28,6 @@ class ServerCombinedErrorGeneratorTest {
         @error("client")
         @retryable
         structure InvalidGreeting {
-            @required
             message: String,
         }
 
@@ -46,54 +45,48 @@ class ServerCombinedErrorGeneratorTest {
         structure Deprecated { }
     """.asSmithyModel()
     private val model = OperationNormalizer.transform(baseModel)
-    private val symbolProvider = serverTestSymbolProvider(model)
+    private val symbolProvider = testSymbolProvider(model)
 
     @Test
     fun `generates combined error enums`() {
         val project = TestWorkspace.testProject(symbolProvider)
         project.withModule(ErrorsModule) {
             listOf("FooException", "ComplexError", "InvalidGreeting", "Deprecated").forEach {
-                model.lookup<StructureShape>("error#$it").serverRenderWithModelBuilder(model, symbolProvider, this)
+                model.lookup<StructureShape>("error#$it").renderWithModelBuilder(model, symbolProvider, this)
             }
             val errors = listOf("FooException", "ComplexError", "InvalidGreeting").map { model.lookup<StructureShape>("error#$it") }
-            ServerCombinedErrorGenerator(
-                model,
-                symbolProvider,
-                symbolProvider.toSymbol(model.lookup("error#Greeting")),
-                errors,
-            ).render(this)
+            val generator = OperationErrorGenerator(model, symbolProvider, symbolProvider.toSymbol(model.lookup("error#Greeting")), errors)
+            generator.render(this)
 
             unitTest(
                 name = "generates_combined_error_enums",
                 test = """
-                    let variant = InvalidGreeting { message: String::from("an error") };
-                    assert_eq!(format!("{}", variant), "InvalidGreeting: an error");
-                    assert_eq!(variant.message(), "an error");
-                    assert_eq!(
-                        variant.retryable_error_kind(),
-                        aws_smithy_types::retry::ErrorKind::ClientError
-                    );
-
-                    let error = GreetingError::InvalidGreeting(variant);
+                    let kind = GreetingErrorKind::InvalidGreeting(InvalidGreeting::builder().message("an error").build());
+                    let error = GreetingError::new(kind, aws_smithy_types::Error::builder().code("InvalidGreeting").message("an error").build());
+                    assert_eq!(format!("{}", error), "InvalidGreeting: an error");
+                    assert_eq!(error.message(), Some("an error"));
+                    assert_eq!(error.code(), Some("InvalidGreeting"));
+                    use aws_smithy_types::retry::ProvideErrorKind;
+                    assert_eq!(error.retryable_error_kind(), Some(aws_smithy_types::retry::ErrorKind::ClientError));
 
                     // Generate is_xyz methods for errors.
                     assert_eq!(error.is_invalid_greeting(), true);
                     assert_eq!(error.is_complex_error(), false);
 
+                    // Unhandled variants properly delegate message.
+                    let error = GreetingError::generic(aws_smithy_types::Error::builder().message("hello").build());
+                    assert_eq!(error.message(), Some("hello"));
+
+                    let error = GreetingError::unhandled("some other error");
+                    assert_eq!(error.message(), None);
+                    assert_eq!(error.code(), None);
+
                     // Indicate the original name in the display output.
-                    let error = FooException::builder().build();
-                    assert_eq!(format!("{}", error), "FooException");
+                    let error = FooError::builder().build();
+                    assert_eq!(format!("{}", error), "FooError [FooException]");
 
                     let error = Deprecated::builder().build();
                     assert_eq!(error.to_string(), "Deprecated");
-                """,
-            )
-
-            unitTest(
-                name = "generates_converters_into_combined_error_enums",
-                test = """
-                    let variant = InvalidGreeting { message: String::from("an error") };
-                    let error: GreetingError = variant.into();
                 """,
             )
